requirements.txt
pillow
qrcode
ascon
opencv-python
tabulate
streamlit>=1.28.0